rename zlib to zlibfreeimage
