+++ Source/FreeImage/PluginPNG.cpp
@@ -40,7 +40,7 @@
 
 // ----------------------------------------------------------
 
-#include "../ZLib/zlib.h"
+#include "../ZLibFreeImage/zlib.h"
 #include "../LibPNG/png.h"
 
 // ----------------------------------------------------------
