+++ Source/ZLibFreeImage/ChangeLog
