+++ Source/ZLibFreeImage/README
