+++ Source/ZLibFreeImage/ZLib.2013.vcxproj
