+++ Source/ZLibFreeImage/ZLib.2017.vcxproj
