+++ Source/ZLibFreeImage/adler32.c
