+++ Source/ZLibFreeImage/compress.c
