+++ Source/ZLibFreeImage/configure
