+++ Source/ZLibFreeImage/crc32.c
