+++ Source/ZLibFreeImage/crc32.h
