+++ Source/ZLibFreeImage/gzclose.c
