+++ Source/ZLibFreeImage/gzguts.h
