+++ Source/ZLibFreeImage/gzlib.c
