+++ Source/ZLibFreeImage/gzwrite.c
