+++ Source/ZLibFreeImage/infback.c
