+++ Source/ZLibFreeImage/inffast.c
