+++ Source/ZLibFreeImage/inffast.h
