+++ Source/ZLibFreeImage/inffixed.h
