+++ Source/ZLibFreeImage/inftrees.h
