+++ Source/ZLibFreeImage/zlib.h
