+++ Source/ZLibFreeImage/zutil.h
